Re-enable shrink test
